+++ e2e/runner/setup_zevm.go
@@ -32,8 +32,6 @@ var EmissionsPoolFunding = big.NewInt(0).Mul(big.NewInt(1e18), big.NewInt(2e7))
 
 // SetTSSAddresses set TSS addresses from information queried from ZetaChain
 func (r *E2ERunner) SetTSSAddresses() error {
-	r.Logger.Print("⚙️ setting up TSS address")
-
 	btcChainID, err := chains.GetBTCChainIDFromChainParams(r.BitcoinParams)
 	if err != nil {
 		return err
@@ -168,6 +166,27 @@ func (r *E2ERunner) SetupTONZRC20() {
 	require.NoError(r, err)
 
 	r.TONZRC20 = TONZRC20
+}
+
+// SetupSUIZRC20 sets up the SUI ZRC20 in the runner from the values queried from the chain
+func (r *E2ERunner) SetupSUIZRC20() {
+	chainID := chains.SuiLocalnet.ChainId
+
+	// noop
+	if r.skipChainOperations(chainID) {
+		return
+	}
+
+	SUIZRC20Addr, err := r.SystemContract.GasCoinZRC20ByChainId(&bind.CallOpts{}, big.NewInt(chainID))
+	require.NoError(r, err)
+
+	r.SUIZRC20Addr = SUIZRC20Addr
+	r.Logger.Info("SUI ZRC20 address: %s", SUIZRC20Addr.Hex())
+
+	SUIZRC20, err := zrc20.NewZRC20(SUIZRC20Addr, r.ZEVMClient)
+	require.NoError(r, err)
+
+	r.SUIZRC20 = SUIZRC20
 }
 
 // EnableHeaderVerification enables the header verification for the given chain IDs