+++ zetaclient/chains/evm/observer/inbound.go
@@ -146,23 +146,28 @@ func (ob *Observer) observeInboundInBlockRange(ctx context.Context, startBlock, 
 		Str(logs.FieldMethod, "observeInboundInBlockRange").
 		Uint64("start_block", startBlock).Uint64("to_block", toBlock).Logger()
 
-	// task 3: query the incoming tx to TSS address (read at most 100 blocks in one go)
-	lastScannedTssRecvd, err := ob.observeTSSReceive(ctx, startBlock, toBlock)
-	if err != nil {
-		logger.Error().Err(err).Msg("error observing TSS received gas asset")
-	}
-
-	// task 4: filter the outbounds from TSS address to supplement outbound trackers
-	// TODO: make this a separate go routine in outbound.go after switching to smart contract V2
-	ob.filterTSSOutbound(ctx, startBlock, toBlock)
-
 	var (
+		lastScannedTssRecvd              = toBlock
 		lastScannedZetaSent              = startBlock - 1
 		lastScannedDeposited             = startBlock - 1
 		lastScannedGatewayDeposit        = startBlock - 1
 		lastScannedGatewayCall           = startBlock - 1
 		lastScannedGatewayDepositAndCall = startBlock - 1
+		err                              error
 	)
+
+	// we now only take these actions on specific configurable chains
+	if !ob.ChainParams().DisableTssBlockScan {
+		// query the incoming tx to TSS address (read at most 100 blocks in one go)
+		lastScannedTssRecvd, err = ob.observeTSSReceive(ctx, startBlock, toBlock)
+		if err != nil {
+			logger.Error().Err(err).Msg("error observing TSS received gas asset")
+		}
+
+		// filter the outbounds from TSS address to supplement outbound trackers
+		// TODO: make this a separate go routine in outbound.go after switching to smart contract V2
+		ob.filterTSSOutbound(ctx, startBlock, toBlock)
+	}
 
 	logs, err := ob.fetchLogs(ctx, startBlock, toBlock)
 	if err != nil {