+++ zetaclient/chains/sui/observer/observer.go
@@ -126,45 +126,16 @@ func (ob *Observer) setLatestGasPrice(price uint64) {
 }
 
 // ensureCursor ensures tx scroll cursor for inbound observations
-func (ob *Observer) ensureCursor(ctx context.Context) error {
+func (ob *Observer) ensureCursor() {
 	if ob.LastTxScanned() != "" {
-		return nil
+		return
 	}
 
 	// Note that this would only work for the empty chain database
 	envValue := os.Getenv(base.EnvVarLatestTxByChain(ob.Chain()))
 	if envValue != "" {
 		ob.WithLastTxScanned(envValue)
-		return nil
 	}
-
-	// let's take the first tx that was ever registered for the Gateway (deployment tx)
-	// Note that this might have for a non-archival node
-	req := models.SuiGetObjectRequest{
-		ObjectId: ob.gateway.PackageID(),
-		Options: models.SuiObjectDataOptions{
-			ShowPreviousTransaction: true,
-		},
-	}
-
-	res, err := ob.client.SuiGetObject(ctx, req)
-	switch {
-	case err != nil:
-		return errors.Wrap(err, "unable to get object")
-	case res.Error != nil:
-		return errors.Errorf("get object error: %s (code %s)", res.Error.Error, res.Error.Code)
-	case res.Data == nil:
-		return errors.New("object data is empty")
-	case res.Data.PreviousTransaction == "":
-		return errors.New("previous transaction is empty")
-	}
-
-	cursor := client.EncodeCursor(models.EventId{
-		TxDigest: res.Data.PreviousTransaction,
-		EventSeq: "0",
-	})
-
-	return ob.setCursor(cursor)
 }
 
 func (ob *Observer) getCursor() string { return ob.LastTxScanned() }